Fix GC3Dintptr and GC3Dsizeiptr typedefs for Win64 https://bugs.webkit.org/show_bug.cgi?id=104426
Patch by Jonathan Liu <net147@gmail.com> on 2012-12-10
Reviewed by Simon Hausmann.

The GC3Dintptr and GC3Dsizeiptr typedefs are incorrect for Win64 as
LLP64 is used there. This would result in compile error due to
narrowing conversion of 64-bit pointer to signed long int which is
32-bit.

* platform/graphics/GraphicsTypes3D.h:

Canonical link: https://commits.webkit.org/122688@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137129 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/GraphicsTypes3D.h
@@ -42,8 +42,8 @@ typedef int GC3Dsizei;
 typedef unsigned int GC3Duint;
 typedef float GC3Dfloat;
 typedef float GC3Dclampf;
-typedef signed long int GC3Dintptr;
-typedef signed long int GC3Dsizeiptr;
+typedef intptr_t GC3Dintptr;
+typedef intptr_t GC3Dsizeiptr;
 typedef char GC3Dchar;
 
 typedef GC3Duint Platform3DObject;
